redesign to file management
moving to a symbolic link-centric approach rather than a file-copying
approach. This entails renaming a bunch of files from dot_filename to
.filename and removing the file-copying scripts.

A few new scripts have been added to improve my workflow (a better
understanding of the purpose of these files has helped tremendously at
reworking them as well).

nvim has been removed from this repo once again. this is part
of an effort to keep this repo focused on smaller tasks. neovim is a
beast of its own and deserved its own repo.